2022 day 18

+++ AdventOfCode/src/aocutil/IO.java
@@ -15,26 +15,34 @@ import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 public class IO {
-	
-	private static final String SESSION;
-	static {
-		final File f = new File("txt/session.txt");
-		if (!f.exists() || !f.isFile()) throw new RuntimeException("Please save your session token to " + f.getAbsolutePath());
-		try (Scanner scan = new Scanner(f)) {
-			SESSION = scan.next();
-		} catch (FileNotFoundException e) {
-			throw new RuntimeException(e);
-		}
-	}
-	
+		
+	/**
+	 * Scan over the problem input
+	 * @param year
+	 * @param day
+	 * @return
+	 */
 	public static Scanner scanner(int year, int day) {
 		return new Scanner(string(year, day));
 	}
 	
+	/**
+	 * Get the entire problem input in a single String
+	 * @param year
+	 * @param day
+	 * @return
+	 */
 	public static String string(int year, int day) {
 		return lines(year, day).collect(Collectors.joining(System.lineSeparator()));
 	}
 	
+	/**
+	 * Iterate over the problem input line by line, splitting each line
+	 * @param year
+	 * @param day
+	 * @param split
+	 * @return
+	 */
 	public static Iterable<String[]> linesIterSplit(int year, int day, String split) {
 		return new Iterable<String[]>() {
 			public Iterator<String[]> iterator() {
@@ -43,6 +51,12 @@ public class IO {
 		};
 	}
 	
+	/**
+	 * Iterate over the problem input line by line
+	 * @param year
+	 * @param day
+	 * @return
+	 */
 	public static Iterable<String> linesIter(int year, int day) {
 		return new Iterable<String>() {
 			public Iterator<String> iterator() {
@@ -52,6 +66,15 @@ public class IO {
 		};
 	}
 	
+	/**
+	 * Stream the problem input line by line
+	 * 
+	 * Base method for other public methods in IO
+	 * 
+	 * @param year
+	 * @param day
+	 * @return
+	 */
 	public static Stream<String> lines(int year, int day) {
 		try (BufferedReader br = new BufferedReader(new FileReader(pullFile(year, day)))) {
 			return br.lines().toList().stream();
@@ -60,6 +83,13 @@ public class IO {
 		}
 	}
 	
+	/**
+	 * Gets a problem input by year and day, either from the local filesystem if it exists, else from AOC
+	 * @param year
+	 * @param day
+	 * @return handle to a file containing the problem input for the given year and day
+	 * @throws IOException
+	 */
 	private static File pullFile(int year, int day) throws IOException {
 		File d = new File("txt/" + year + "/");
 		if (d.exists()) {
@@ -77,21 +107,33 @@ public class IO {
 		return f;
 	}
 	
+	/**
+	 * Blindly downloads the specified URL to the specified file
+	 * @param file where to download to
+	 * @param fileURL where to download from
+	 * @throws IOException uh oh
+	 */
 	private static void downloadFile(File file, String fileURL) throws IOException {
+		final String SESSION;
+		final File f = new File("txt/session.txt");
+		if (!f.exists() || !f.isFile()) throw new RuntimeException("Please save your session token to " + f.getAbsolutePath());
+		try (Scanner scan = new Scanner(f)) {
+			SESSION = scan.next();
+		} catch (FileNotFoundException e) {
+			throw new RuntimeException(e);
+		}
+		
 		final HttpURLConnection httpConn = (HttpURLConnection) new URL(fileURL).openConnection();
 		httpConn.setRequestProperty("Cookie", "session=" + SESSION);
 		final int responseCode = httpConn.getResponseCode();
 
 		// always check HTTP response code first
 		if (responseCode == HttpURLConnection.HTTP_OK) {
-
-//			System.err.println("Content-Type = " + httpConn.getContentType());
-//			System.err.println("Content-Length = " + httpConn.getContentLength());
-
+			
 			// opens input stream from the HTTP connection
-			InputStream inputStream = httpConn.getInputStream();
+			final InputStream inputStream = httpConn.getInputStream();
 			// opens an output stream to save into file
-			FileOutputStream outputStream = new FileOutputStream(file);
+			final FileOutputStream outputStream = new FileOutputStream(file);
 
 			int bytesRead = -1;
 			byte[] buffer = new byte[4096];
